requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
groq==0.11.0
tavily-python==0.3.9
httpx==0.28.0
pydantic==2.10.1
python-dotenv==1.0.1
aiosqlite==0.20.0
sse-starlette==2.1.3
python-multipart==0.0.12
PyGithub==2.5.0
markdown==3.7
omium
google-genai
